+++ flake.nix
@@ -9,8 +9,8 @@
       forAllSystems' = nixpkgs.lib.genAttrs;
       forAllSystems = forAllSystems' supportedSystems;
 
-      raylibRev = "f5328a9bb63a0e0eca7dead15cfa01a3ec1417c2";
-      raylibHash = "sha256-gAv1jfMdbKWnlIqqBddVn+WE0BOIARMmagpK61bxVnU=";
+      raylibRev = "2a0963ce0936abe0cd3ec32882638d860e435d16";
+      raylibHash = "sha256-4p3nq04irS8AFojH88Bh1r8KiOjQhZf7nFmQhf1EDU8=";
       rayguiRev = "1e03efca48c50c5ea4b4a053d5bf04bad58d3e43";
       rayguiHash = "sha256-PzQZxCz63EPd7sVFBYY0T1s9jA5kOAxF9K4ojRoIMz4=";